// src/app/open-api/model/participantType.ts
/**
 * kkl Education API
 * Education new project|
 *
 * The version of the OpenAPI document: 1.11.1
 * Contact: yochai@one1.co.il
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ParticipantType { 
    id?: number;
    name?: string;
}